continus move of hGraphQL files
