--- Projects/App/GraphQL/Giraffe/MyHome.graphql
@@ -1,15 +1,0 @@
-query MyHome {
-    insurance {
-        address
-        postalNumber
-        type
-        livingSpace
-        ancillaryArea
-        yearOfConstruction
-        numberOfBathrooms
-        extraBuildings {
-            ...ExtraBuildingFragment
-        }
-        isSubleted
-    }
-}
