+++ Projects/Home/GraphQL/Giraffe/ContractStatusFragment.graphql
